Wrote some tests, discovered fatal flaw in impl.
The paper http://hwml.com/LeibRamp.pdf cannot be converted to work with
ints, as the final step is multiplying or dividing prev_delay by a
number. With ints, that number can be 0, 1, or >1. If 0, movement stops.
If 1, movement stays the same, no acceleration. If >1, movement at least
doubles, leading to at least exponential accelleration.

Need to find/develop a new algo.